refactor: 게임 재시작, 종료 선택하는 메서드의 주체를 야구게임에서 플레이어로 변경
